+++ feature/home/src/main/java/com/dev/james/booktracker/home/presentation/screens/ReadGoalsScreen.kt
@@ -1,28 +1,16 @@
 package com.dev.james.booktracker.home.presentation.screens
 
 import android.widget.Toast
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.FastOutSlowInEasing
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInHorizontally
-import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutHorizontally
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -30,24 +18,12 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ElevatedButton
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExposedDropdownMenuBox
-import androidx.compose.material3.ExposedDropdownMenuDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -55,44 +31,33 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.layoutId
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.constraintlayout.compose.ConstraintLayout
-import androidx.constraintlayout.compose.ConstraintSet
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.dev.james.booktracker.compose_ui.ui.components.RoundedBrownButton
 import com.dev.james.booktracker.compose_ui.ui.components.StandardToolBar
-import com.dev.james.booktracker.compose_ui.ui.components.WeekDaySelectorComponent
 import com.dev.james.booktracker.compose_ui.ui.theme.BookAppTypography
 import com.dev.james.booktracker.home.R
+import com.dev.james.booktracker.home.presentation.components.BottomNextPreviousButtons
+import com.dev.james.booktracker.home.presentation.components.TextFieldComponent
+import com.dev.james.booktracker.home.presentation.forms.CurrentReadForm
+import com.dev.james.booktracker.home.presentation.forms.OverallGoalsForm
 import com.dev.james.booktracker.home.presentation.viewmodels.ImageSelectorUiState
 import com.dev.james.booktracker.home.presentation.viewmodels.ReadGoalsScreenViewModel
 import com.dev.james.booktracker.home.presentation.navigation.HomeNavigator
 import com.dev.james.booktracker.home.presentation.viewmodels.ReadGoalsScreenState
 import com.dsc.form_builder.BaseState
-import com.dsc.form_builder.ChoiceState
 import com.dsc.form_builder.FormState
-import com.dsc.form_builder.SelectState
 import com.dsc.form_builder.TextFieldState
 import com.ramcosta.composedestinations.annotation.Destination
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import timber.log.Timber
 
 @Composable
 @Destination
@@ -320,716 +285,5 @@ fun StatelessReadGoalScreen(
     }
 }
 
-@Composable
-@Preview("CurrentReadForm")
-fun CurrentReadForm(
-    //will take in form state
-    currentReadFormState: FormState<TextFieldState> = FormState(fields = listOf()),
-    imageSelectorState: ImageSelectorUiState = ImageSelectorUiState(),
-    onSaveBookClicked: () -> Unit = {},
-    imageSelectorClicked: () -> Unit = {}
-) {
-
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = Modifier.padding(16.dp)
-    ) {
-
-        ImageSelectorComponent(
-            onSelect = {
-                //start the image picker
-                imageSelectorClicked()
-            },
-            imageSelectorState = imageSelectorState
-        )
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-
-        val titleFieldState: TextFieldState = currentReadFormState.getState(name = "title")
-
-        TextFieldComponent(
-            modifier = Modifier.fillMaxWidth(),
-            label = "Title",
-            text = titleFieldState.value,
-            hasError = titleFieldState.hasError,
-            onTextChanged = { fieldValue ->
-                titleFieldState.change(fieldValue)
-            }
-        )
-
-        Spacer(
-            modifier = Modifier
-                .height(16.dp)
-                .padding(start = 16.dp)
-        )
-
-        val authorFieldState: TextFieldState = currentReadFormState.getState("author")
-        TextFieldComponent(
-            modifier = Modifier.fillMaxWidth(),
-            label = "Author",
-            text = authorFieldState.value,
-            hasError = authorFieldState.hasError,
-            onTextChanged = { fieldValue ->
-                authorFieldState.change(fieldValue)
-            }
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        val chapterDropDownState: TextFieldState = currentReadFormState.getState("chapters")
-        val currentChapterDropDownState: TextFieldState =
-            currentReadFormState.getState("current chapter")
-
-        Row(
-            horizontalArrangement = Arrangement.SpaceEvenly,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-
-            val chapterCount = mutableListOf<String>()
-            for (chapter in 1..50) {
-                chapterCount.add(
-                    chapter.toString()
-                )
-            }
-
-
-
-            DropDownComponent(
-                modifier = Modifier.width(180.dp),
-                label = "Chapters",
-                selectedText = chapterDropDownState.value,
-                dropDownItems = chapterCount.toList(),
-                hasError = chapterDropDownState.hasError,
-                onListItemSelected = { chapter ->
-                    Timber.tag("ReadGoalsScreen").d(chapter)
-                    chapterDropDownState.change(chapter)
-                    //update chapter count
-                },
-                canUserFill = true
-            )
-
-            Spacer(modifier = Modifier.width(30.dp))
-
-
-
-            DropDownComponent(
-                modifier = Modifier.width(180.dp),
-                label = "Current chapter",
-                selectedText = currentChapterDropDownState.value,
-                dropDownItems = chapterCount.toList(),
-                hasError = currentChapterDropDownState.hasError,
-                onListItemSelected = { chapter ->
-                    Timber.tag("ReadGoalsScreen").d(chapter)
-                    currentChapterDropDownState.change(chapter)
-                    //update current selected chapter
-                },
-                canUserFill = true
-            )
-
-        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        val chapterTitleState: TextFieldState = currentReadFormState.getState("chapter title")
-
-        TextFieldComponent(
-            modifier = Modifier.fillMaxWidth(),
-            text = chapterTitleState.value,
-            label = "Current chapter title",
-            hasError = chapterTitleState.hasError,
-            onTextChanged = { text ->
-                chapterTitleState.change(text)
-            }
-        )
-
-        Spacer(modifier = Modifier.height(12.dp))
-
-        ElevatedButton(
-            modifier = Modifier.fillMaxWidth(),
-            onClick = {
-                //validate with state show error where necessary
-                onSaveBookClicked()
-
-            },
-            shape = RoundedCornerShape(10.dp),
-            contentPadding = PaddingValues(16.dp),
-            colors = ButtonDefaults.buttonColors(
-                containerColor = MaterialTheme.colorScheme.secondary
-            )
-        ) {
-            Text(
-                text = "Add book",
-                style = BookAppTypography.labelMedium,
-                textAlign = TextAlign.Center,
-                color = MaterialTheme.colorScheme.onPrimary
-            )
-        }
-
-    }
-
-}
-
-@Composable
-@Preview("ImageSelectorComponent")
-fun ImageSelectorComponent(
-    imageSelectorState: ImageSelectorUiState = ImageSelectorUiState(),
-    onSelect: () -> Unit = {}
-) {
-
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier
-            .width(158.dp)
-            .height(191.dp)
-            .clip(shape = RoundedCornerShape(20.dp))
-            .background(color = MaterialTheme.colorScheme.secondaryContainer)
-            .border(
-                width = if (imageSelectorState.isError) 2.dp else 0.dp,
-                color = if (imageSelectorState.isError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.secondaryContainer,
-                shape = RoundedCornerShape(20.dp)
-            )
-            .clickable {
-                //starts the image picking flow
-                onSelect()
-            }
-    ) {
-
-        if (imageSelectorState.imageSelectedUri.isNotBlank()) {
-            //will replace with coil
-            Image(
-                painterResource(id = R.drawable.image_placeholder_24),
-                contentDescription = "taken book image",
-                modifier = Modifier
-                    .width(80.dp)
-                    .height(80.dp),
-                contentScale = ContentScale.Fit
-            )
-        }
-
-        if (imageSelectorState.imageSelectedUri.isBlank() && !imageSelectorState.showProgress) {
-            Column(
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Icon(
-                    painterResource(id = R.drawable.add_a_photo_24),
-                    contentDescription = "Add image placeholder",
-                    modifier = Modifier
-                        .height(80.dp)
-                        .width(80.dp),
-                    tint = MaterialTheme.colorScheme.secondary
-
-                )
-                Text(
-                    text = if (imageSelectorState.isError) "Add image*" else "Add image",
-                    style = BookAppTypography.labelLarge,
-                    modifier = Modifier.padding(16.dp),
-                    textAlign = TextAlign.Center,
-                    color = if (imageSelectorState.isError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.secondary
-                )
-            }
-        }
-
-        if (imageSelectorState.showProgress) {
-            CircularProgressIndicator(
-                color = MaterialTheme.colorScheme.secondary
-            )
-        }
-
-    }
-
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-@Preview(name = "TextFieldComponent")
-fun TextFieldComponent(
-    modifier: Modifier = Modifier,
-    text: String = "",
-    hasError: Boolean = false,
-    label: String = "",
-    onTextChanged: (String) -> Unit = {}
-) {
-    //var currentText by remember { mutableStateOf("") }
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.Start,
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        if (label.isNotBlank()) {
-            Text(
-                text = if (hasError) "$label*" else label,
-                style = BookAppTypography.labelMedium,
-                color = if (hasError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.primary
-            )
-        }
-
-        OutlinedTextField(
-            value = text,
-            onValueChange = {
-                onTextChanged(it)
-            },
-            modifier = modifier
-                .padding(top = 8.dp)
-                .border(
-                    width = 2.dp,
-                    shape = RoundedCornerShape(0.dp),
-                    //if error is available , change the color of border to error color
-                    color = if (hasError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.secondary
-                ),
-
-            colors = TextFieldDefaults.textFieldColors(
-                containerColor = MaterialTheme.colorScheme.background,
-                errorCursorColor = MaterialTheme.colorScheme.error
-            ),
-            textStyle = BookAppTypography.bodyMedium,
-            shape = RoundedCornerShape(0.dp),
-            //adjust error depending if error is available
-            isError = hasError,
-        )
-
-    }
-
-}
-
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalComposeUiApi::class)
-@Composable
-@Preview("DropDownComponent")
-fun DropDownComponent(
-    modifier: Modifier = Modifier,
-    selectedText: String = "",
-    hasError: Boolean = false,
-    label: String = "",
-    placeHolderText: String = "",
-    dropDownItems: List<String> = listOf(),
-    onListItemSelected: (String) -> Unit = {},
-    canUserFill: Boolean = false
-) {
-    /*
-    component for a dropdown menu
-    */
-
-    var isExpanded by remember { mutableStateOf(false) }
-    val keyboardController = LocalSoftwareKeyboardController.current
-
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.Start
-    ) {
-
-        if (label.isNotBlank()) {
-            Text(
-                modifier = if (label.length > 15) Modifier.fillMaxWidth() else Modifier,
-                text = if (hasError) "$label*" else label,
-                style = BookAppTypography.labelMedium,
-                color = if (hasError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.primary,
-            )
-        }
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        ExposedDropdownMenuBox(
-            expanded = isExpanded,
-            onExpandedChange = {
-                /*if (!canUserFill) {
-                    keyboardController?.hide()
-                }*/
-                isExpanded = it
-            },
-            modifier = modifier
-        ) {
-
-            TextField(
-                value = selectedText.ifBlank { placeHolderText },
-                readOnly = !canUserFill,
-                onValueChange = {
-                    if (canUserFill) {
-                        onListItemSelected(it)
-                    }
-                },
-                modifier = modifier
-                    .border(
-                        width = 2.dp,
-                        shape = RoundedCornerShape(0.dp),
-                        color = if (hasError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.secondary
-                    )
-                    .menuAnchor(),
-                colors = TextFieldDefaults.textFieldColors(
-                    containerColor = MaterialTheme.colorScheme.background,
-                    cursorColor = if (canUserFill) MaterialTheme.colorScheme.secondary else Color.Transparent
-                ),
-                shape = RoundedCornerShape(0.dp),
-                trailingIcon = {
-                    ExposedDropdownMenuDefaults.TrailingIcon(expanded = isExpanded)
-                },
-                keyboardOptions = KeyboardOptions(
-                    keyboardType = KeyboardType.Number
-                ),
-                keyboardActions = KeyboardActions(
-                    onDone = { keyboardController?.hide() }
-                ),
-                singleLine = true
-
-            )
-
-            ExposedDropdownMenu(
-                expanded = isExpanded,
-                onDismissRequest = { isExpanded = false }
-            ) {
-                dropDownItems.forEach { item ->
-                    DropdownMenuItem(
-                        onClick = {
-                            onListItemSelected(item)
-                            isExpanded = false
-                        },
-                        text = {
-                            Text(text = item, style = BookAppTypography.bodySmall)
-                        }
-                    )
-                }
-            }
-
-        }
-    }
-}
-
-
-@Composable
-@Preview("BottomButtonsSection", showBackground = true)
-fun BottomNextPreviousButtons(
-    modifier: Modifier = Modifier,
-    onNextClicked: () -> Unit = {},
-    onPreviousClicked: () -> Unit = {},
-    currentPosition: Int = 0
-) {
-    Box(
-        modifier = modifier,
-        contentAlignment = Alignment.BottomCenter
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = if (currentPosition > 0) Arrangement.SpaceBetween else Arrangement.End
-        ) {
-
-            AnimatedVisibility(
-                visible = currentPosition > 0,
-                enter = fadeIn(
-                    animationSpec = tween(
-                        durationMillis = 100,
-                        delayMillis = 10,
-                        easing = FastOutSlowInEasing
-                    )
-                ),
-                exit = fadeOut(
-                    animationSpec = tween(
-                        durationMillis = 100,
-                        delayMillis = 10,
-                        easing = FastOutSlowInEasing
-                    )
-                )
-            ) {
-                RoundedBrownButton(
-                    label = "Previous",
-                    cornerRadius = 8.dp
-                ) {
-                    onPreviousClicked()
-                }
-            }
-
-            RoundedBrownButton(
-                label = if (currentPosition == 1) "Finish" else "Next",
-                cornerRadius = 8.dp
-            ) {
-                onNextClicked()
-            }
-
-        }
-
-
-    }
-
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-@Preview(name = "OverallGoalsForm", showBackground = true)
-fun OverallGoalsForm(
-    modifier: Modifier = Modifier,
-    overallGoalsFormState: FormState<BaseState<out Any>> = FormState(fields = listOf()),
-    alertSwitchState: Boolean = false,
-    onAlertSwitchChecked: (Boolean) -> Unit = {},
-    startDialPickerDialog: () -> Unit = {},
-    startAlarmPickerDialog: () -> Unit = {}
-) {
-
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = Modifier.padding(start = 16.dp, end = 16.dp)
-    ) {
-
-        Text(
-            text = "Overall",
-            modifier = Modifier.fillMaxWidth(),
-            style = BookAppTypography.headlineMedium,
-            color = MaterialTheme.colorScheme.primary
-        )
-
-        Spacer(
-            modifier = Modifier
-                .height(16.dp)
-                .fillMaxWidth()
-        )
-
-        val minTimeField = overallGoalsFormState.getState<TextFieldState>(name = "time")
-        Text(
-            text = "In a day , I want to read for at least.",
-            modifier = Modifier.fillMaxWidth(),
-            style = BookAppTypography.labelMedium,
-            color = if (minTimeField.hasError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.primary,
-        )
-
-        Row(
-            horizontalArrangement = Arrangement.SpaceEvenly,
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.padding(top = 8.dp)
-        ) {
-
-            OutlinedTextField(
-                value = minTimeField.value,
-                onValueChange = {
-                    minTimeField.change(it)
-                },
-                modifier = Modifier
-                    .padding(top = 8.dp)
-                    .border(
-                        width = 2.dp,
-                        shape = RoundedCornerShape(0.dp),
-                        //if error is available , change the color of border to error color
-                        color = if (minTimeField.hasError) MaterialTheme.colorScheme.error else MaterialTheme.colorScheme.secondary
-                    )
-                    .weight(1f),
-
-                colors = TextFieldDefaults.textFieldColors(
-                    containerColor = MaterialTheme.colorScheme.background,
-                    errorCursorColor = MaterialTheme.colorScheme.error
-                ),
-                textStyle = BookAppTypography.bodyMedium,
-                shape = RoundedCornerShape(0.dp),
-                //adjust error depending if error is available
-                isError = minTimeField.hasError,
-            )
-
-            Spacer(modifier = Modifier.width(20.dp))
-
-            RoundedBrownButton(
-                label = "Set",
-                icon = R.drawable.baseline_access_alarm_24,
-                onClick = {
-                    //launch alarm picker dialog
-                },
-                cornerRadius = 10.dp
-            )
-        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        val dropDownItems = listOf<String>(
-            "Every day",
-            "Every day except",
-            "Weekend only",
-            "Weekdays",
-            "Select specific days"
-        )
-
-        val frequencyDropDownState =
-            overallGoalsFormState.getState<ChoiceState>(name = "frequency field")
-
-        //will show whether or not if specific days are shown
-        val supportedDays = listOf(
-            "Select specific days" ,  "Every day except"
-        )
-        val daysSelectorState = overallGoalsFormState.getState<SelectState>(name = "specific days")
-
-
-       DropDownComponent(
-                label = "For how long do you want this goal to run?",
-                modifier = Modifier.fillMaxWidth(),
-                selectedText = frequencyDropDownState.value,
-                canUserFill = false,
-                dropDownItems = dropDownItems,
-                onListItemSelected = {
-                    //show items
-                    frequencyDropDownState.change(it)
-                },
-                placeHolderText = "",
-                hasError = frequencyDropDownState.hasError
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-         AnimatedVisibility(
-                visible = supportedDays.contains(frequencyDropDownState.value) ,
-                exit = slideOutVertically(
-                    animationSpec = tween(durationMillis = 500) ,
-                    targetOffsetY = { -it }
-                ) ,
-                enter = slideInVertically(
-                    animationSpec = tween(durationMillis = 500) ,
-                    initialOffsetY = { -it }
-                )
-            ) {
-                WeekDaySelectorComponent(
-                    modifier = Modifier.fillMaxWidth(),
-                    onDaySelected = { day ->
-                        //update selected days list
-                        if (daysSelectorState.value.contains(day)) {
-                            daysSelectorState.unselect(day)
-                        } else {
-                            daysSelectorState.select(day)
-                        }
-                    },
-                    selectedDays = daysSelectorState.value
-                )
-            }
-
-
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        SwitchSection(
-            isChecked = alertSwitchState,
-            onCheckChanged = { status ->
-                //update checked state
-                onAlertSwitchChecked(status)
-            }
-        )
-
-        //show this section whether or when the switch is on or off
-        val noteFieldState = overallGoalsFormState.getState<TextFieldState>("alert note")
-        AlertSetupSection(
-            noteText = noteFieldState.value,
-            onNoteFieldChange = { note ->
-                noteFieldState.change(note)
-            }
-        )
-
-
-    }
-
-
-}
-
-@Composable
-fun AlertSetupSection(
-    noteText: String = "",
-    hasError: Boolean = false,
-    timeSet: String = "",
-    onNoteFieldChange: (String) -> Unit = {},
-    onTimeSelectorClicked: () -> Unit = {}
-) {
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Spacer(modifier = Modifier.height(16.dp))
-
-        TextFieldComponent(
-            modifier = Modifier.fillMaxWidth(),
-            label = "Add note",
-            onTextChanged = { note ->
-                //update text action
-                onNoteFieldChange(note)
-            },
-            text = noteText
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            text = "Time Set",
-            style = BookAppTypography.labelMedium,
-            color = MaterialTheme.colorScheme.primary,
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceEvenly
-        ) {
-            Text(
-                text = "00:00 AM",
-                style = BookAppTypography.headlineLarge,
-                modifier = Modifier.weight(1f),
-                color = MaterialTheme.colorScheme.secondary,
-                fontSize = 32.sp
-            )
-
-            RoundedBrownButton(
-                icon = R.drawable.baseline_access_alarm_24,
-                label = "Set",
-                onClick = {
-                    //open
-                    onTimeSelectorClicked()
-                },
-                cornerRadius = 10.dp
-            )
-
-        }
-
-    }
-}
-
-@Composable
-fun SwitchSection(
-    isChecked: Boolean = false,
-    onCheckChanged: (Boolean) -> Unit = {}
-) {
-
-    Row(
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Text(
-            text = "Do you wish to get alerts?",
-            style = BookAppTypography.bodyMedium,
-            color = MaterialTheme.colorScheme.primary,
-            modifier = Modifier.weight(1f)
-        )
-
-
-        Switch(
-
-            checked = isChecked,
-            onCheckedChange = { status ->
-                //update check state
-                onCheckChanged(status)
-            },
-            thumbContent = {
-                if (isChecked) {
-                    Text(
-                        text = "On",
-                        style = BookAppTypography.labelSmall
-                    )
-                } else {
-                    Text(
-                        text = "Off",
-                        style = BookAppTypography.labelSmall
-                    )
-                }
-
-            }
-
-        )
-
-    }
-
-}
